moved modeltable to model.rs
