+++ gateway/src/function.rs
@@ -4,7 +4,6 @@ use std::collections::HashMap;
 use tracing::instrument;
 use uuid::Uuid;
 
-use crate::config_parser::ModelTable;
 use crate::embeddings::EmbeddingModelConfig;
 use crate::endpoints::inference::InferenceParams;
 use crate::error::{Error, ErrorDetails};
@@ -14,6 +13,7 @@ use crate::inference::types::{
 };
 use crate::jsonschema_util::{JSONSchemaFromPath, JsonSchemaRef};
 use crate::minijinja_util::TemplateConfig;
+use crate::model::ModelTable;
 use crate::tool::{DynamicToolParams, StaticToolConfig, ToolCallConfig, ToolChoice};
 use crate::variant::{InferenceConfig, Variant, VariantConfig};
